port check 128: check_copyright_notice_is_consistent_across_family
(issue #1388)
